+++ UniversalDeviceToolkit.WPF/Controls/Macro/AbstractMacroEventControl.xaml.cs
@@ -4,6 +4,7 @@ using System.Windows.Input;
 using Humanizer;
 using LenovoLegionToolkit.Lib.Extensions;
 using UniversalDeviceToolkit.Lib.Macro;
+using UniversalDeviceToolkit.Lib.Macro.Resources;
 using Wpf.Ui.Controls;
 
 namespace UniversalDeviceToolkit.WPF.Controls.Macro
@@ -37,16 +38,16 @@ public abstract partial class AbstractMacroEventControl
         _header.Title = (macroEvent.Source, macroEvent.Direction, macroEvent.Key) switch
         {
             (MacroSource.Keyboard, _, _) => KeyInterop.KeyFromVirtualKey((int)macroEvent.Key).ToString(),
-            (MacroSource.Mouse, MacroDirection.Move, _) => "MOVE",
-            (MacroSource.Mouse, MacroDirection.Wheel, >= 0x80000000) => "WHEEL DOWN",
-            (MacroSource.Mouse, MacroDirection.Wheel, _) => "WHEEL UP",
-            (MacroSource.Mouse, MacroDirection.HorizontalWheel, >= 0x80000000) => "WHEEL LEFT",
-            (MacroSource.Mouse, MacroDirection.HorizontalWheel, _) => "WHEEL RIGHT",
-            (MacroSource.Mouse, _, >= 0xFF) => "XBUTTON" + (macroEvent.Key >> 16),
-            (MacroSource.Mouse, _, 1) => "LBUTTON",
-            (MacroSource.Mouse, _, 2) => "RBUTTON",
-            (MacroSource.Mouse, _, 3) => "MBUTTON",
-            (MacroSource.Mouse, _, _) => "BUTTON" + macroEvent.Key,
+            (MacroSource.Mouse, MacroDirection.Move, _) => Resource.MacroEventControl_MouseMove_Title,
+            (MacroSource.Mouse, MacroDirection.Wheel, >= 0x80000000) => Resource.MacroEventControl_MouseWheelDown_Title,
+            (MacroSource.Mouse, MacroDirection.Wheel, _) => Resource.MacroEventControl_MouseWheelUp_Title,
+            (MacroSource.Mouse, MacroDirection.HorizontalWheel, >= 0x80000000) => Resource.MacroEventControl_MouseWheelLeft_Title,
+            (MacroSource.Mouse, MacroDirection.HorizontalWheel, _) => Resource.MacroEventControl_MouseWheelRight_Title,
+            (MacroSource.Mouse, _, >= 0xFF) => string.Format(Resource.MacroEventControl_MouseXButton_Title, macroEvent.Key >> 16),
+            (MacroSource.Mouse, _, 1) => Resource.MacroEventControl_MouseLButton_Title,
+            (MacroSource.Mouse, _, 2) => Resource.MacroEventControl_MouseRButton_Title,
+            (MacroSource.Mouse, _, 3) => Resource.MacroEventControl_MouseMButton_Title,
+            (MacroSource.Mouse, _, _) => string.Format(Resource.MacroEventControl_MouseButton_Title, macroEvent.Key),
             _ => string.Empty
         };